refactor: unify transport reconnection logic and move policy to base
