+++ test/unit/transport/reconnect_logic_test.cc
@@ -7,7 +7,7 @@
 
 #include "unilink/config/tcp_client_config.hpp"
 #include "unilink/diagnostics/error_types.hpp"
-#include "unilink/transport/tcp_client/reconnect_policy.hpp"
+#include "unilink/transport/base/reconnect_policy.hpp"
 
 using namespace unilink;
 using namespace unilink::transport;
