+++ test/unit/transport/test_reconnect_decider.cc
@@ -6,7 +6,7 @@
 #include "unilink/config/tcp_client_config.hpp"
 #include "unilink/diagnostics/error_types.hpp"
 #include "unilink/transport/tcp_client/detail/reconnect_decider.hpp"
-#include "unilink/transport/tcp_client/reconnect_policy.hpp"
+#include "unilink/transport/base/reconnect_policy.hpp"
 
 using namespace unilink;
 using namespace unilink::transport::detail;
